Mongo - Fix 502 Error
How to fix the 502 error on the DISA site:
Problem:  the reason we keep running into this issue is because we're accessing a shared server. So it might be possible that someone is accidentally pulling or containers down or the server admin runs regular clean ups which causes the containers to stop running. 
When you're on the Brown network run these commands from your terminal.